+++ tests/conftest.py
@@ -1558,3 +1558,64 @@ def fixture_cg_context(
     cg_config.status_db_ = base_store
     cg_config.housekeeper_api_ = housekeeper_api
     return cg_config
+
+
+@pytest.fixture(name="case_id_with_single_sample")
+def case_id_with_single_sample():
+    """Return a case id that should only be associated with one sample."""
+    return "exhaustedcrocodile"
+
+
+@pytest.fixture(name="case_id_with_multiple_samples")
+def case_id_with_multiple_samples():
+    """Return a case id that should be associated with multiple samples."""
+    return "righteouspanda"
+
+
+@pytest.fixture(name="case_id_without_samples")
+def case_id_without_samples():
+    """Return a case id that should not be associated with any samples."""
+    return "confusedtrout"
+
+
+@pytest.fixture(name="sample_id_in_single_case")
+def sample_id_in_single_case():
+    """Return a sample id that should be associated with a single case."""
+    return "ASM1"
+
+
+@pytest.fixture(name="sample_id_in_multiple_cases")
+def sample_id_in_multiple_cases():
+    """Return a sample id that should be associated with multiple cases."""
+    return "ASM2"
+
+
+@pytest.fixture(name="store_with_multiple_cases_and_samples")
+def store_with_multiple_cases_and_samples(
+    case_id_without_samples: str,
+    case_id_with_single_sample: str,
+    case_id_with_multiple_samples: str,
+    sample_id_in_single_case: str,
+    sample_id_in_multiple_cases: str,
+    case_id: str,
+    helpers: StoreHelpers,
+    store: Store,
+):
+    """Return a store containing multiple cases and samples."""
+
+    helpers.add_case(store=store, internal_id=case_id_without_samples)
+    helpers.add_case_with_samples(
+        base_store=store, case_id=case_id_with_multiple_samples, nr_samples=5
+    )
+
+    case_samples: List[Tuple[str, str]] = [
+        (case_id_with_multiple_samples, sample_id_in_multiple_cases),
+        (case_id, sample_id_in_multiple_cases),
+        (case_id_with_single_sample, sample_id_in_single_case),
+    ]
+
+    for case_sample in case_samples:
+        case_id, sample_id = case_sample
+        helpers.add_case_with_sample(base_store=store, case_id=case_id, sample_id=sample_id)
+
+    yield store
