+++ tests/store/test_status_samples_filters.py
@@ -1,6 +1,8 @@
 from alchy import Query
 
 from cg.constants.subject import PhenotypeStatus
+from cg.store import Store
+from cg.store.models import Sample
 from cg.store.status_sample_filters import (
     get_samples_with_loqusdb_id,
     get_samples_without_loqusdb_id,
@@ -49,3 +51,26 @@ def test_get_samples_without_loqusdb_id(helpers, store, sample_store, sample_id,
     # THEN the obtained sample should match the expected one
     assert sample in not_uploaded_samples
     assert sample_uploaded not in not_uploaded_samples
+
+
+def test_get_sample_by_entry_id(sample_store: Store):
+    """Test retrieving sample by entry id."""
+    # GIVEN a store containing samples
+    # WHEN retrieving a sample by its entry id
+    sample: Sample = sample_store.get_sample_by_id(entry_id=1)
+
+    # THEN a sample should be returned
+    assert sample
+
+
+def test_get_sample_by_internal_id(
+    store_with_multiple_cases_and_samples: Store, sample_id_in_single_case: str
+):
+    # GIVEN a store containing a sample
+    # WHEN retrieving the sample by its internal id
+    sample: Sample = store_with_multiple_cases_and_samples.sample(
+        internal_id=sample_id_in_single_case
+    )
+
+    # THEN it is returned
+    assert sample
